# swagger-output_production.json
{
  "swagger": "2.0",
  "info": {
    "title": "API DOC",
    "description": ""
  },
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "headers",
      "name": "authorization",
      "description": "Token 前請加上 Bearer，<code>Bearer ＜Token＞</code>"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/posts/": {
      "get": {
        "tags": [
          "動態貼文"
        ],
        "summary": "取得所有貼文",
        "description": "取得所有貼文",
        "parameters": [],
        "responses": {
          "200": {
            "description": "取得<strong>全部貼文</strong>",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "62924730a3cca4c1b4a52226"
                },
                "userData": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "62908a8ebf85a52ce75989a5"
                    },
                    "userName": {
                      "type": "string",
                      "example": "jimmyWu"
                    },
                    "email": {
                      "type": "string",
                      "example": "gg@mail.com"
                    },
                    "password": {
                      "type": "string",
                      "example": "123456"
                    },
                    "avatarUrl": {
                      "type": "string",
                      "example": "https://avatars.githubusercontent.com/u/42748910?v=4"
                    }
                  }
                },
                "content": {
                  "type": "string",
                  "example": "test"
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "type": {
                  "type": "string",
                  "example": "person"
                },
                "image": {
                  "type": "string",
                  "example": "http://"
                },
                "likes": {
                  "type": "number",
                  "example": 0
                },
                "comments": {
                  "type": "number",
                  "example": 0
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "動態貼文"
        ],
        "summary": "張貼個人動態",
        "description": "新增單筆貼文",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "type": "object",
            "required": true,
            "description": "資料格式查看必填欄位，點按下方 Model 切換後，屬性欄位名稱的後方紅色的 <span>*</span>",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "example": "test"
                },
                "type": {
                  "type": "string",
                  "example": "person"
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "image": {
                  "type": "string",
                  "example": "http://"
                }
              },
              "required": [
                "content"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "建立貼正正確回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "userData": {
                      "type": "string",
                      "example": "629a3dae000fcec3e68c92c7"
                    },
                    "content": {
                      "type": "string",
                      "example": "test"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "type": {
                      "type": "string",
                      "example": "person"
                    },
                    "image": {
                      "type": "string",
                      "example": "http://"
                    },
                    "likes": {
                      "type": "number",
                      "example": 0
                    },
                    "comments": {
                      "type": "number",
                      "example": 0
                    },
                    "_id": {
                      "type": "string",
                      "example": "629a3fa4f6a449c76689d3ff"
                    },
                    "createAt": {
                      "type": "string",
                      "example": "2022-06-03T17:06:44.297Z"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "動態貼文"
        ],
        "summary": "刪除所有貼文",
        "description": "刪除所有貼文",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  },
...
}